Move tests around sensibly
